requirements.txt
flask>=3.0.3,<3.1.0
flask-login>=0.6.3,<0.7.0
flask-sqlalchemy>=3.1.1,<3.2.0
flask-migrate>=4.0.7,<4.1.0
flask-wtf>=1.2.1,<1.3.0
werkzeug>=3.0.3,<3.1.0
email-validator>=2.2.0,<2.3.0
wtforms>=3.1.2,<3.2.0
psycopg2-binary
pyjwt>=2.8.0,<2.9.0
python-dotenv~=1.0.1
flask-restx
SQLAlchemy~=2.0.31
alembic~=1.13.2
cryptography==43.0.0
flask-cors>=4.0.1,<4.1.1